src: mavlink: Move helper functions to utils.rs
