+++ src/mavlink/mod.rs
@@ -1,2 +1,3 @@
 pub mod manager;
 pub mod mavlink_camera;
+mod utils;
